+++ armsrc/felica.c
@@ -620,7 +620,7 @@ void felica_sniff(uint32_t samplesToSkip, uint32_t triggersToSkip) {
     set_tracelen(numbts);
     set_tracelen(BigBuf_max_traceLen());
 
-    Dbprintf("Felica sniffing done, tracelen: %i, use hf list felica for annotations", BigBuf_get_traceLen());
+    Dbprintf("Felica sniffing done, tracelen: %i, use " _YELLOW_("`hf felica list`") " for annotations", BigBuf_get_traceLen());
     reply_mix(CMD_ACK, 1, numbts, 0, 0, 0);
     LED_D_OFF();
 }