+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project uses the changelog in accordance with [keepchangelog](http://keepac
  - Fix 'hf iclass wrbl' - dealing with tags in unsecured vs secured pagemode now is correct (@iceman1001)
  - Change many commands to cliparser (@iceman1001, @tcprst, @mwalker33,...)
  - ...
+ - Change `hf iclass chk/lookup/loclass` speedups (@iceman1001)
+ - Change - ongoing convertion to cliparser (@tcprst, @iceman1001)
  - Added compilation options for 256k Proxmark versions, see doc (@doegox)
  - Added support for 10b UID in `hf 14a sim` (@doegox)
  - Added `HF_TCPRST` standalone mode which read and emulate IKEA Rothult cards (@tcprst)